[fakeintake] move to fakeintake
