+++ test/fakeintake/docker-compose.yaml
