+++ test/new-e2e/utils/payloadtesthelper/fixtures/README.md
@@ -1,3 +1,3 @@
 # Fixtures
 
-Fixtures are curl dump from requests to a fake-intake running on ddev. Used in unit test to validate payload parsing works as expected.
+Fixtures are curl dump from requests to a fakeintake running on ddev. Used in unit test to validate payload parsing works as expected.
